spi: cadence-quadspi: Implement Suspend Resume
The cqspi suspend and resume functions were earlier not working
correctly when attempting to suspend the device. These changes are tested
on a TI AM62x SK EVM. Prior to this fix suspend-resume used
to fail in this qspi driver. This caused the data read from the flash
after the resume to be corrupted and unusable.

The fix has now been verified after comparing data written to the flash
and reading it back, both before and after system suspend-resume and made
sure it all matches.

Signed-off-by: Dhruva Gole <d-gole@ti.com>