firmware: ti_sci: Add support for querying the firmware caps
This patch adds support for the TISCI_MSG_QUERY_FW_CAPS
message, used to retrieve the firmware capabilities of the
currently running system firmware. The message belongs to
the TISCI general core message API [1] and is available in
SysFW version 08.04.03 and above. Currently, the message is
supported on devices with split architecture of the system
firmware (DM + TIFS) like AM62x. Old revisions or not yet
supported platforms will NACK this request.

We're using this message locally in ti_sci.c to get the low
power featutes of the FW/SoC. As there's no other kernel
consumers yet, this is not added to struct ti_sci_core_ops.

Also have a has_lpm list of SOCs to ensure that proper LPM
capabilities are used and if any versions of firmware that happen to
advertise garbage value for fw_caps are ignored.
This was needed as the API is buggy on TI SYSFW v09.00 and below

[1] https://software-dl.ti.com/tisci/esd/latest/2_tisci_msgs/general/core.html

Signed-off-by: Georgi Vlaev <g-vlaev@ti.com>
[d-gole@ti.com: add workaround for sysfw fw_caps inconsistency]
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Message-ID: <20230804115037.754994-3-d-gole@ti.com>